Move all ui classes into a .ui package
