+++ code/app/src/main/java/com/kernelcrew/moodapp/ui/MainActivity.java
@@ -1,4 +1,4 @@
-package com.kernelcrew.moodapp;
+package com.kernelcrew.moodapp.ui;
 
 import android.os.Bundle;
 
@@ -7,6 +7,8 @@ import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.graphics.Insets;
 import androidx.core.view.ViewCompat;
 import androidx.core.view.WindowInsetsCompat;
+
+import com.kernelcrew.moodapp.R;
 
 public class MainActivity extends AppCompatActivity {
     @Override
