+++ code/app/src/main/res/layout/fragment_auth_sign_in.xml
@@ -5,7 +5,7 @@
     android:id="@+id/container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AuthSignIn">
+    tools:context=".ui.AuthSignIn">
 
     <com.google.android.material.appbar.AppBarLayout
         android:id="@+id/appBarLayout"
