+++ code/app/src/main/res/layout/fragment_auth_sign_up.xml
@@ -4,7 +4,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AuthSignUp">
+    tools:context=".ui.AuthSignUp">
 
     <com.google.android.material.appbar.AppBarLayout
         android:id="@+id/appBarLayout"
