+++ code/app/src/main/res/layout/fragment_home_feed.xml
@@ -5,7 +5,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".HomeFeed">
+    tools:context=".ui.HomeFeed">
 
     <com.google.android.material.appbar.AppBarLayout
         android:id="@+id/appBarLayout2"
