+++ code/app/src/main/res/layout/fragment_my_profile.xml
@@ -5,7 +5,7 @@ xmlns:tools="http://schemas.android.com/tools"
 xmlns:app="http://schemas.android.com/apk/res-auto"
 android:layout_width="match_parent"
 android:layout_height="match_parent"
-tools:context=".MyProfile">
+tools:context=".ui.MyProfile">
 
 <com.google.android.material.appbar.AppBarLayout
     android:id="@+id/appBarLayout2"
